refactor: 기존 WinningParser를 더 포괄적인 클래스명으로 변경
